pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-polling"
version = "0.1.0"
description = "AI Public Opinion Polling Data Extraction Pipeline"
authors = [
    {name = "AI Polling Team", email = "noetel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Researchers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "tenacity>=8.0.0",
    "openpyxl>=3.1.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "python-dateutil>=2.8.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ai-polling = "ai_polling.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/ai-polling"
Repository = "https://github.com/your-org/ai-polling"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_polling*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true